+++ packages/ses/index.js
@@ -12,7 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { globalThis, Error, assign } from './src/commons.js';
+/* global process */
+
+import {
+  globalThis,
+  Error,
+  assign,
+  defineProperty,
+  getOwnPropertyDescriptor,
+} from './src/commons.js';
 import { tameFunctionToString } from './src/tame-function-tostring.js';
 import { getGlobalIntrinsics } from './src/intrinsics.js';
 import { getAnonymousIntrinsics } from './src/get-anonymous-intrinsics.js';
@@ -30,6 +38,28 @@ function getThis() {
 
 if (getThis()) {
   throw new Error(`SES failed to initialize, sloppy mode (SES_NO_SLOPPY)`);
+}
+
+// Protect against the hazard presented by Node.js domains.
+if (typeof process === 'object' && process !== null) {
+  // Check whether domains were initialized.
+  const domainDescriptor = getOwnPropertyDescriptor(process, 'domain');
+  if (domainDescriptor !== undefined && domainDescriptor.get !== undefined) {
+    throw new Error(
+      `SES failed to initialized, Node.js domains have been initialized (SES_NO_DOMAINS)`,
+    );
+  }
+  // Prevent domains from initializing.
+  // This is clunky because the exception thrown from the domains package does
+  // not direct the user's gaze toward a knowledge base about the problem.
+  // The domain module merely throws an exception when it attempts to define
+  // the domain property of the process global during its initialization.
+  defineProperty(process, 'domain', {
+    value: null,
+    configurable: false,
+    writable: false,
+    enumerable: false,
+  });
 }
 
 const markVirtualizedNativeFunction = tameFunctionToString();